#31 ブラウザの更新ボタンを押した際にstateが初期化される問題に対応
