+++ src/components/pages/LoginPage/index.tsx
@@ -1,7 +1,12 @@
 import { FC } from 'react';
 
 import LoginTemplate from '../../templates/LoginTemlate';
+import useLoggedInState from '../../../hooks/useLoggedInState';
 
-const Index: FC = () => <LoginTemplate />;
+const Index: FC = () => {
+  useLoggedInState();
+
+  return <LoginTemplate />;
+};
 
 export default Index;
