+++ src/components/pages/StoreManagementPage/index.tsx
@@ -1,33 +1,10 @@
-import { FC, useContext, useEffect } from 'react';
+import { FC } from 'react';
+import useLoggedInState from '../../../hooks/useLoggedInState';
 
 import StoreManagementTemplate from '../../templates/StoreManagementTemplate';
-import { AuthContext } from '../../../contexts/Auth';
-import createAxiosClient from '../../../api/client';
-
-type UserIdResponseType = {
-  data: {
-    userId: string;
-  };
-};
 
 const Index: FC = () => {
-  const { setLoggedInUserId } = useContext(AuthContext);
-  const axiosClient = createAxiosClient();
-
-  useEffect(() => {
-    axiosClient('check/user_id')
-      .then((res: UserIdResponseType) => {
-        if (res.data.userId === 'none') {
-          setLoggedInUserId(undefined);
-        } else {
-          setLoggedInUserId(res.data.userId);
-        }
-      })
-      .catch((err) => {
-        // eslint-disable-next-line
-        console.log(err);
-      });
-  });
+  useLoggedInState();
 
   return (
     <>
